requirements.txt
pandas
requests
bs4
transformers
scikit-learn
tensorboardX
#gradio
schedule
tqdm
pyspellchecker
paramiko
pymysql
sshtunnel
#pip install torch==1.12.1+cu116 torchvision==0.13.1+cu116 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu116
#pip install torch==1.13.1+cu117 --extra-index-url https://download.pytorch.org/whl/cu117
